refactor(openomni): split runtime binding responsibilities
